src/ingestion.py
import pandas as pd
from sqlalchemy import create_engine
from dotenv import load_dotenv
import os

load_dotenv()

DB_HOST = os.getenv('DB_HOST')
DB_PORT = os.getenv('DB_PORT')
DB_NAME = os.getenv('DB_NAME')
DB_USER = os.getenv('DB_USER')
DB_PASS = os.getenv('DB_PASS')
DB_SCHEMA = os.getenv('DB_SCHEMA')

DATABASE_URL = f"postgresql://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"

engine = create_engine(DATABASE_URL)
print(engine)

df_raw = pd.read_csv('../raw/dados_ficha_a_desafio.csv')
print(df_raw.shape)
df_raw.to_sql('raw_data', engine, if_exists='replace', index=False, chunksize=1000)
print('Dados inseridos com sucesso!')